Incluido a conexão com banco de dados e ajustado o CRUD

+++ 04_restWithASPNET10Study/restWithASPNET10Study/restWithASPNET10Study/Services/IUserService.cs
@@ -6,10 +6,10 @@ namespace restWithASPNET10Study.Services
     public interface IUserService
     {
         User Create(User user);
-        User FindById(int id);
+        User FindById(long id);
         List<User> FindAll();
         User Update(User user);
-        void Delete(int id);
+        void Delete(long id);
 
     }
 }
